feat: Integrate user profile, experience, pref loc, and certif on one dummy page
