+++ src/main.tsx
@@ -7,7 +7,6 @@ import { PreviewTalentModule } from './modules/PreviewTalentModule';
 import { RegisterModule } from './modules/RegisterFormModule';
 import LoginModule from './modules/LoginFormModule';
 import { TalentProfilePage } from './modules/TalentProfileModule';
-import { TalentProfilePage } from './modules/TestCertifModule';
 
 createRoot(document.getElementById('root')!).render(
   <BrowserRouter>
@@ -20,7 +19,6 @@ createRoot(document.getElementById('root')!).render(
         <Route path="/register" element={<RegisterModule />} />
         <Route path="/login" element={<LoginModule />} />
         <Route path="/me" element={<TalentProfilePage />} />
-        <Route path="/test" element={<TalentProfilePage />} />
       </Route>
     </Routes>
   </BrowserRouter>
