+++ src/modules/TalentProfileModule/index.tsx
@@ -1,12 +1,12 @@
 import { Typography } from '@/components';
-import { TalentExperience } from './Section/talentExperience';
+import { TalentInformation } from './Section/talentInformation';
 
 
 export const TalentProfilePage = () => {
   return (
     <>
         <Typography variant="h1">Talent Profile</Typography>
-        <TalentExperience />
+        <TalentInformation />
     </>
   );
 };
