--- src/modules/TestCertifModule/Section/talentCertificate.tsx
@@ -1,56 +1,0 @@
-import { Button } from '@/components';
-import Certificate, { CertificateDetail } from '@/components/ui/certificate';
-import { ArrowLeft, BookmarkIcon } from 'lucide-react';
-
-export const TalentCertificate: React.FC = () => {
-
-    const certificate: CertificateDetail[] = [
-        {
-          id: 1,
-          title: 'Sertifikasi 1',
-          file: new File(['12345'], "Sertifikasi 1.pdf", { type: "application/pdf" })
-        },
-        {
-          id: 2,
-          title: 'Sertifikasi 2',
-          file: new File(['1234'], "Sertifikasi 2.pdf", { type: "application/pdf" })
-        },
-        {
-          id: 3,
-          title: 'Sertifikasi 3',
-          file: new File(["dummy content"], "Sertifikasi 3.pdf", { type: "application/pdf" })
-        },
-      ];
-      
-
-  return !certificate ? (
-    <div className="absolute inset-0 flex h-full w-full items-center justify-center">
-      <div
-        data-testid="loading-spinner"
-        className="h-8 w-8 animate-spin rounded-full border-4 border-gray-300 border-t-blue-500"
-      ></div>
-    </div>
-  ) : (
-    <div className="flex w-full justify-center">
-      <div className="m-6 w-full max-w-6xl justify-center bg-white">
-        <div className="flex w-full justify-between p-4">
-          <Button variant="primary-outline" className="flex py-2">
-            <ArrowLeft size={20} />
-            <span>Kembali</span>
-          </Button>
-          <Button variant="primary" className="flex py-2">
-            <BookmarkIcon className="h-6 w-6 text-white" />
-            <span>Masukkan ke Favorit</span>
-          </Button>
-        </div>
-
-        <div className="flex w-full flex-col items-center space-y-2 p-4 md:flex-row md:items-start md:space-x-6 md:space-y-0">
-          <img src="image-3.png" alt="Logo" className="h-[330px] w-[298px]" />
-          <div className="w-full flex-col items-center space-y-4">
-            <Certificate certificates={certificate} />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
