--- src/modules/TestCertifModule/index.tsx
@@ -1,12 +1,0 @@
-import { Typography } from '@/components';
-import { TalentCertificate } from './Section/talentCertificate';
-
-
-export const TalentProfilePage = () => {
-  return (
-    <>
-        <Typography variant="h1">Talent Profile</Typography>
-        <TalentCertificate />
-    </>
-  );
-};
